lesson1/requirements.txt
openai
langchain
langchain-openai
langchain-community
duckduckgo-search
crewai
python-dotenv